[refactor] move getReadableDate to utils
