+++ Frontend/src/app/components/roaster-revision/roaster-revision.component.ts
@@ -12,6 +12,7 @@ import {
   getCoordinatesDiffParts,
   getMultilineDiffParts,
 } from '../../utils/diff.utils';
+import { getReadableDate } from '../../utils/date.utils';
 
 @Component({
   selector: 'app-roaster-revision',
@@ -171,19 +172,7 @@ export class RoasterRevisionComponent implements OnInit {
   }
 
   getReadableDate(dateIsoString: string | null): string | null {
-    if (!dateIsoString) return null;
-    const date = new Date(dateIsoString);
-
-    const year = date.getFullYear();
-    const month = String(date.getMonth() + 1).padStart(2, '0');
-    const day = String(date.getDate()).padStart(2, '0');
-
-    let hours = date.getHours();
-    const ampm = hours >= 12 ? 'PM' : 'AM';
-    hours = hours % 12 || 12; // convert to 12-hour format
-    const minutes = String(date.getMinutes()).padStart(2, '0');
-
-    return `${year}-${month}-${day} at ${hours}:${minutes} ${ampm}`;
+    return getReadableDate(dateIsoString);
   }
 
   public getMultilineDiffParts(
